requirements.txt
psycopg2
django
Pillow
django-durationwidget